fix:  확정 페이지에 미참가자가 진입하는 경우 막는 로직 추가
